make sure

+++ wifi.cpp
@@ -45,7 +45,7 @@ void setupWifi() {
  WiFi.macAddress(mac);
  uint8_t check = 0;
  for ( int i = 0 ; i < 6 ; i++ ) {
-  check += mac[i];
+  check += (uint8_t)mac[i];
  }
  sprintf((char*)state.uid,"%02X%02X%02X%02X%02X%02X%02X",mac[0],mac[1],mac[2],mac[3],mac[4],mac[5],check);
  LOGF(("UID : %s\r\n",(char*)state.uid));
